backend/requirements.txt
# Web Framework
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-cors==4.0.0
flask-mail==0.9.1
flask-jwt-extended==4.5.3

# Authentication & Security
authlib==1.2.1
pyjwt==2.8.0
cryptography==41.0.4
pyotp==2.8.0
bcrypt==4.0.1
webauthn==1.11.1

# Crypto
web3==6.11.0
ecdsa==0.18.0
bitcoinlib==0.7.5

# Utilities
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
python-multipart==0.0.6

# Database
alembic==1.12.1
psycopg2-binary==2.9.7
